refactor: consolidate and simplify management commands for chemical and soil hardness data import
